pyproject.toml
[tool.poetry]
name = "airbnb-challenge"
version = "0.1.0"
description = ""
authors = ["rafy <juvwind@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.1"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
